refactor(mcp): decompose monolithic server into modular architecture
Break down the 3,237-line agent_builder_server.py into focused modules
following the single-responsibility principle. Each handler module
registers its tools on the shared FastMCP instance via register(mcp).

New structure:
- session.py: BuildSession model and persistence
- handlers/session_handler.py: Session CRUD tools
- handlers/goal_handler.py: Goal definition tools
- handlers/graph_handler.py: Node/edge CRUD tools
- handlers/export_handler.py: Graph export + README generation
- handlers/mcp_server_handler.py: External MCP server management
- handlers/simulation_handler.py: Node/graph simulation tools
- handlers/evaluation_handler.py: Evaluation rules + plan tools
- handlers/test_generation_handler.py: Test guideline generation
- handlers/test_execution_handler.py: Test run/debug/list tools
- validation/validators.py: Graph validation + credential checks

All files are under 500 lines. Public API remains backwards-compatible
via re-exports in agent_builder_server.py.

Closes #1880
