+++ core/framework/mcp/handlers/__init__.py
@@ -1,0 +1,5 @@
+"""
+MCP tool handler modules for the Agent Builder server.
+
+Each module registers its tools on a shared FastMCP instance.
+"""
