+++ core/framework/mcp/session.py
@@ -1,0 +1,168 @@
+"""
+Session management for the MCP Agent Builder.
+
+Handles session creation, persistence, loading, and lifecycle management.
+Sessions store the in-progress agent build state including goal, nodes, edges,
+and MCP server configurations.
+"""
+
+import json
+from datetime import datetime
+from pathlib import Path
+
+from framework.graph import Constraint, EdgeCondition, EdgeSpec, Goal, NodeSpec, SuccessCriterion
+
+# Session persistence directory
+SESSIONS_DIR = Path(".agent-builder-sessions")
+ACTIVE_SESSION_FILE = SESSIONS_DIR / ".active"
+
+
+class BuildSession:
+    """Build session with persistence support."""
+
+    def __init__(self, name: str, session_id: str | None = None):
+        self.id = session_id or f"build_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
+        self.name = name
+        self.goal: Goal | None = None
+        self.nodes: list[NodeSpec] = []
+        self.edges: list[EdgeSpec] = []
+        self.mcp_servers: list[dict] = []
+        self.created_at = datetime.now().isoformat()
+        self.last_modified = datetime.now().isoformat()
+
+    def to_dict(self) -> dict:
+        """Serialize session to dictionary."""
+        return {
+            "session_id": self.id,
+            "name": self.name,
+            "goal": self.goal.model_dump() if self.goal else None,
+            "nodes": [n.model_dump() for n in self.nodes],
+            "edges": [e.model_dump() for e in self.edges],
+            "mcp_servers": self.mcp_servers,
+            "created_at": self.created_at,
+            "last_modified": self.last_modified,
+        }
+
+    @classmethod
+    def from_dict(cls, data: dict) -> "BuildSession":
+        """Deserialize session from dictionary."""
+        session = cls(name=data["name"], session_id=data["session_id"])
+        session.created_at = data.get("created_at", session.created_at)
+        session.last_modified = data.get("last_modified", session.last_modified)
+
+        # Restore goal
+        if data.get("goal"):
+            goal_data = data["goal"]
+            session.goal = Goal(
+                id=goal_data["id"],
+                name=goal_data["name"],
+                description=goal_data["description"],
+                success_criteria=[
+                    SuccessCriterion(**sc) for sc in goal_data.get("success_criteria", [])
+                ],
+                constraints=[Constraint(**c) for c in goal_data.get("constraints", [])],
+            )
+
+        # Restore nodes
+        session.nodes = [NodeSpec(**n) for n in data.get("nodes", [])]
+
+        # Restore edges
+        edges_data = data.get("edges", [])
+        for e in edges_data:
+            # Convert condition string back to enum
+            condition_str = e.get("condition")
+            if isinstance(condition_str, str):
+                condition_map = {
+                    "always": EdgeCondition.ALWAYS,
+                    "on_success": EdgeCondition.ON_SUCCESS,
+                    "on_failure": EdgeCondition.ON_FAILURE,
+                    "conditional": EdgeCondition.CONDITIONAL,
+                    "llm_decide": EdgeCondition.LLM_DECIDE,
+                }
+                e["condition"] = condition_map.get(condition_str, EdgeCondition.ON_SUCCESS)
+            session.edges.append(EdgeSpec(**e))
+
+        # Restore MCP servers
+        session.mcp_servers = data.get("mcp_servers", [])
+
+        return session
+
+
+# Global session state
+_session: BuildSession | None = None
+
+
+def _ensure_sessions_dir():
+    """Ensure sessions directory exists."""
+    SESSIONS_DIR.mkdir(exist_ok=True)
+
+
+def save_session(session: BuildSession):
+    """Save session to disk."""
+    _ensure_sessions_dir()
+
+    # Update last modified
+    session.last_modified = datetime.now().isoformat()
+
+    # Save session file
+    session_file = SESSIONS_DIR / f"{session.id}.json"
+    with open(session_file, "w") as f:
+        json.dump(session.to_dict(), f, indent=2, default=str)
+
+    # Update active session pointer
+    with open(ACTIVE_SESSION_FILE, "w") as f:
+        f.write(session.id)
+
+
+def _load_session(session_id: str) -> BuildSession:
+    """Load session from disk."""
+    session_file = SESSIONS_DIR / f"{session_id}.json"
+    if not session_file.exists():
+        raise ValueError(f"Session '{session_id}' not found")
+
+    with open(session_file) as f:
+        data = json.load(f)
+
+    return BuildSession.from_dict(data)
+
+
+def _load_active_session() -> BuildSession | None:
+    """Load the active session if one exists."""
+    if not ACTIVE_SESSION_FILE.exists():
+        return None
+
+    try:
+        with open(ACTIVE_SESSION_FILE) as f:
+            session_id = f.read().strip()
+
+        if session_id:
+            return _load_session(session_id)
+    except Exception:
+        pass
+
+    return None
+
+
+def get_session() -> BuildSession:
+    """Get current session, loading from disk if needed."""
+    global _session
+
+    # Try to load active session if no session in memory
+    if _session is None:
+        _session = _load_active_session()
+
+    if _session is None:
+        raise ValueError("No active session. Call create_session first.")
+
+    return _session
+
+
+def set_current_session(session: BuildSession | None):
+    """Set the current in-memory session."""
+    global _session
+    _session = session
+
+
+def get_current_session_raw() -> BuildSession | None:
+    """Get the current session without raising if None."""
+    return _session
