+++ core/framework/mcp/validation/__init__.py
@@ -1,0 +1,5 @@
+"""
+Validation modules for the MCP Agent Builder.
+
+Provides graph validation, context flow analysis, and credential checking.
+"""
